feat: move browser ui logic into a context
